### 2023-10-27 #135 @whedon
[PRE REVIEW]: FlowFPX: Nimble Tools for Debugging Floating-Point Exceptions
**Submitting author:** @ashton314 (<a href="http://orcid.org/0000-0001-5524-7930">Ashton Wiersdorf</a>)
**Repository:** https://github.com/utahplt/juliacon2023-paper
**Branch with paper.md** (empty if default branch):
**Version:**
**Managing EiC:** Mathieu Besançon
