### 2023-10-27 #135 @lucaferranti
looking at the error

```
===========Latexmk: Missing input file: 'cleveref.sty' from line
  '! LaTeX Error: File `cleveref.sty' not found.'
Latexmk: Missing input file: 'cleveref.sty' from line
  '! LaTeX Error: File `cleveref.sty' not found.'
Failure to make 'paper.pdf'
      Refer to 'paper.log' for details
Looks like we failed to compile the PDF
```

it seems like a missing library file from the repository
