### 2023-10-27 #135 @ashton314
I see—thanks for that! I was frustrated with the Whedon bot because I couldn't figure out how to view the compile errors. Is there a place where I can see those?

I think `cleverref.sty` is part of my LaTeX distro; I'll add it manually if I can.
