### 2023-10-27 2f22fad
Add some DOIs that Whedon found
