+++ ref.bib
@@ -186,6 +186,7 @@ year = {2016}
   ISSN = {1570-2820},
   MRCLASS = {65Y15},
   MRNUMBER = {3043640},
+  DOI = {10.1515/jnum-2012-0013},
   URL = {https://freefem.org/}
 }
 
@@ -206,6 +207,7 @@ year = {2019}
   volume={111},
   pages={102870},
   year={2022},
+  doi={10.1016/j.parco.2021.102870},
   publisher={Elsevier}
 }
 
