### 2023-10-27 #135 @ashton314
DOIs:

 - `10.2307/2317055` this is a review of The Art of Computer Programming, not the book itself. I'm having difficulty finding a DOI for the second volume; we have the ISBN in the bibtex file; can we ask the citation generation format to put the ISBN in instead?
 - `10.1515/jnum-2012-0013` added
 - `10.1016/j.parco.2021.102870` added
