### 2023-10-27 #135 @ashton314
Alright, @lucaferranti I'm a little stuck on this: I'm not sure what this error is saying. Have you seen this before?
