### 2023-10-27 #135 @lucaferranti
> 10.2307/2317055 this is a review of The Art of Computer Programming, not the book itself. I'm having difficulty finding a DOI for the second volume; we have the ISBN in the bibtex file; can we ask the citation generation format to put the ISBN in instead?

The references check is not perfect and can have some false-positives sometimes (like in this case). Great that you double-checked! If you find no DOI, it's ok to leave it this way.

> LaTeX error

ouch, those can be a bit challenging sometimes.

are you able to compile the paper locally with `latexmk` ? Does your paper rely on any fancy library / font installed in your system? If yes, those should be included in the source code too, since on CI we only use plain `latexmk`.

I have not seen that specific error before, but it seems to indicate that some file (maybe some font) is missing.

Googling your error, I found this: https://www.mrunix.de/forums/showthread.php?61988-pdflatex-psyr-und-pcrr7t-not-found (discussion is in German unfortunately, but google translate apparently can translate the page pretty well), which seems to indeed to refer to some missing fonts.

You don't import many packages in the paper

```latex
\usepackage{amsmath}
\usepackage{cleveref}
\usepackage{microtype}
\usepackage{pifont}
```

and `amsmath` is unlikely to be the issue. So some of those packages is needing something that latexmk does not have by default. Given the error mentions some fonts, my uneducated guess would be to start with `pifont`, but not sure. Could you try if removing that (and the parts in the paper that rely on it) fixes the error? If it works, then we found the issue, otherwise we can do some more debugging.

Here a quick debugging I could get done on Friday night, hopefully it helps. Let me know how this proceeds. :)
