### 2023-10-27 #135 @ashton314
> references check

Well, it was very helpful with the other two missing DOIs that it found!

I can build the paper locally with `latexmk`, as can the rest of my team. I'll see what exactly the `pifont` package is doing. Someone else added that.

(I actually speak German, so that page was very helpful! :)
