### 2023-10-27 #135 @ashton314
Looks like the `pifont` package makes the postscript dingbats accessible. My guess is the font files that it needs are not installed wherever Whedon runs.

Do you know of a way to add fonts? I suppose I could just drop the font file in the repo but that seems a little… overkill. I don't really like the default fonts anyway… I'd love it if I could use [JuliaMono](https://juliamono.netlify.app/) for the code samples—it'd be appropriate, no?
