### 2023-10-27 #135 @ashton314
Ok, I've found a workaround for the missing fonts; we just needed an "x" and a checkmark character from the dingbats; I've hacked around that. Now something's wrong with the citations. I can take a look at this more later, but maybe you'd be able to see something right away that's wrong?
