### 2023-10-29 #135 @lucaferranti
Oh gosh, it seems like it does not build the bibilography, since the references don't work.

I am not sure why that happens, fwiw, I could build the paper locally.

I see you have a `Makefile` with content

```
paper.pdf: paper.tex ref.bib
	latexmk -bibtex -pdf paper.tex

.PHONY: clean
clean:
	latexmk -c && rm paper.bbl
```

I am not sure the `paper.bbl` should be removed though :thinking:

This is just a weak hunch, but maybe Whedon automatically detects the `Makefile` (which does not come with the template), and runs that, which messes up the bot workflow. Could try

1. Removing the `rm paper.bbl` OR
2. Removing the `Makefile`

I know it's a pretty weak guess, but it's the only thing I can think of rn.

(we are in the process of updating the infrastructure, and the process should become much smoother in the foreseable future, but we are not quite there yet, sorry for the inconvenience)
