### 2023-10-30 #135 @ashton314
@lucaferranti I've removed the Makefile from the repo (well, moved it to something *not* called Makefile) and Whedon still isn't happy. The `.bbl` file isn't checked into our repo anyway.
