### 2023-10-30 #135 @lucaferranti
oh my, I'll try to have a closer look. The process should be much smoother (or at least easier to debug) once we complete the infrastructure update, let's see which one I get done first :)

Some random ideas:

- I see from the error message that on CI `latexmk` v 4.55 is used? Which version are you using locally to build the paper? Maybe the older version does not support some formatting in the `.bib` file?
