### 2023-10-30 #135 @ashton314
```
➤ ~ latexmk --version
Latexmk, John Collins, 7 Jan. 2023. Version 4.79
```

I'm using what looks like a pretty up-to-date version. I think the most recent version is `4.80`. Maybe I can try building with the older version of latexmk, but it will take me a minute to figure out how to install that older version.

Thanks so much for your help Luca!
