### 2023-11-07 #135 @lucaferranti
A possible strategy to debug could be to comment out most references from the bibtex file and see if it works with a minimal set and gradually reinsert others. My suspect is that the older version does not support some bibtex features and this could help identify the problematic entries.

I am sorry for the incovenience with this, we are upgrading `whedon` to the new `editorial-bot` and that should fix all of these issues. Moreover, the new version will use `texlive-full` which should be fully battery included. The upgrade is going a bit slow, but getting there soon. You can either wait for us to upgrade the infrastructure or try to get latex to work with the old bot. I hope to wrap up the upgrade within a couple of weeks.
