+++ ref.bib
@@ -1,33 +1,22 @@
 @article{goldbergWhatEveryComputer1991,
   title = {What Every Computer Scientist Should Know about Floating-Point Arithmetic},
   author = {Goldberg, David},
-  year = {1991},
   journal = {ACM Computing Surveys},
   volume = {23},
   number = {1},
   pages = {5--48},
-  issn = {0360-0300, 1557-7341},
   doi = {10.1145/103162.103163},
-  url = {https://dl.acm.org/doi/10.1145/103162.103163},
-  urldate = {2023-04-21},
-  abstract = {Floating-point arithmetic is considered as esoteric subject by many people. This is rather surprising, because floating-point is ubiquitous in computer systems: Almost every language has a floating-point datatype; computers from PCs to supercomputers have floating-point accelerators; most compilers will be called upon to compile floating-point algorithms from time to time; and virtually every operating system must respond to floating-point exceptions such as overflow. This paper presents a tutorial on the aspects of floating-point that have a direct impact on designers of computer systems. It begins with background on floating-point representation and rounding error, continues with a discussion of the IEEE floating point standard, and concludes with examples of how computer system builders can better support floating point.},
-  langid = {english}
+  year = {1991}
 }
 
-@incollection{heislerFinchDomainSpecific2022,
-  title = {Finch: {{Domain Specific Language}} and {{Code Generation}} for {{Finite Element}} and {{Finite Volume}} in {{Julia}}},
+@inproceedings{heislerFinchDomainSpecific2022,
+  title = {{Finch}: {{Domain Specific Language}} and {{Code Generation}} for {{Finite Element}} and {{Finite Volume}} in {{Julia}}},
   booktitle = {{ICCS}},
   author = {Heisler, Eric and Deshmukh, Aadesh and Sundar, Hari},
-  year = {2022},
-  volume = {13350},
   pages = {118--132},
   publisher = {{Springer}},
   doi = {10.1007/978-3-031-08751-6\_9},
-  url = {https://link.springer.com/10.1007/978-3-031-08751-6_9},
-  urldate = {2023-01-27},
-  isbn = {978-3-031-08750-9 978-3-031-08751-6},
-  langid = {english},
-  file = {/Users/ashton/Zotero/storage/MKS3PE2A/Heisler et al. - 2022 - Finch Domain Specific Language and Code Generatio.pdf}
+  year = {2022}
 }
 
 @article{IEEEStandardBinary1985,
@@ -42,117 +31,94 @@
 @misc{p-draft-1997,
   title={Handling {IEEE 754} Invalid Operation Exceptions in Real Interval Arithmetic},
   author={Priest, David J.},
-  journal={Manuscript, mai},
-  note={\url{https://j3-fortran.org/doc/year/97/97-172.pdf}. Accessed 2023-08-16},
+  note={\url{https://j3-fortran.org/doc/year/97/97-172.pdf}, Accessed 2023-08-16},
   year={1997}
 }
 
 @phdthesis{klowerLowprecisionClimateComputing2021,
-  type = {{{http://purl.org/dc/dcmitype/Text}}},
   title = {Low-Precision Climate Computing: Preserving Information despite Fewer Bits},
   author = {Kl{\"o}wer, Milan},
-  year = {2021},
   url = {https://ora.ox.ac.uk/objects/uuid:1158e44a-7faf-45a0-8ab1-73c91fd694a6},
   urldate = {2023-05-12},
[…]
