+++ _ref.bib
@@ -1,0 +1,369 @@
+@article{goldbergWhatEveryComputer1991,
+  title = {What Every Computer Scientist Should Know about Floating-Point Arithmetic},
+  author = {Goldberg, David},
+  journal = {ACM Computing Surveys},
+  volume = {23},
+  number = {1},
+  pages = {5--48},
+  doi = {10.1145/103162.103163},
+  year = {1991}
+}
+
+@inproceedings{heislerFinchDomainSpecific2022,
+  title = {{Finch}: {{Domain Specific Language}} and {{Code Generation}} for {{Finite Element}} and {{Finite Volume}} in {{Julia}}},
+  booktitle = {{ICCS}},
+  author = {Heisler, Eric and Deshmukh, Aadesh and Sundar, Hari},
+  pages = {118--132},
+  publisher = {{Springer}},
+  doi = {10.1007/978-3-031-08751-6\_9},
+  year = {2022}
+}
+
+@article{IEEEStandardBinary1985,
+  title = {{{IEEE Standard}} for {{Binary Floating-Point Arithmetic}}},
+  key = {{IEEE}},
+  journal = {{ANSI/IEEE Std 754-1985}},
+  pages = {1--20},
+  doi = {10.1109/IEEESTD.1985.82928},
+  year = {1985}
+}
+
+@misc{p-draft-1997,
+  title={Handling {IEEE 754} Invalid Operation Exceptions in Real Interval Arithmetic},
+  author={Priest, David J.},
+  note={\url{https://j3-fortran.org/doc/year/97/97-172.pdf}, Accessed 2023-08-16},
+  year={1997}
+}
+
+@phdthesis{klowerLowprecisionClimateComputing2021,
+  title = {Low-Precision Climate Computing: Preserving Information despite Fewer Bits},
+  author = {Kl{\"o}wer, Milan},
+  url = {https://ora.ox.ac.uk/objects/uuid:1158e44a-7faf-45a0-8ab1-73c91fd694a6},
+  urldate = {2023-05-12},
+  school = {University of Oxford},
+  year = {2021}
+}
+
+@article{klowerNumberFormatsError2020,
+  title = {Number {{Formats}}, {{Error Mitigation}}, and {{Scope}} for 16-{{Bit Arithmetics}} in {{Weather}} and {{Climate Modeling Analyzed With}} a {{Shallow Water Model}}},
+  author = {Kl{\"o}wer, Milan and D{\"u}ben, Peter D. and Palmer, T. N.},
+  journal = {Journal of Advances in Modeling Earth Systems},
+  volume = {12},
+  number = {10},
+  doi = {10.1029/2020MS002246},
+  year = {2020}
+}
+
+@inproceedings{klowerPositsAlternativeFloats2019,
+  title = {Posits as an Alternative to Floats for Weather and Climate Models},
+  booktitle = {{{Conference}} for {{Next Generation Arithmetic}}},
+  author = {Kl{\"o}wer, Milan and D{\"u}ben, Peter D. and Palmer, Tim N.},
+  pages = {1--8},
+  publisher = {{ACM}},
+  doi = {10.1145/3316279.3316281},
+  year = {2019}
+}
+
+@book{knuthArtComputerProgramming1997,
+  title = {The {{Art}} of {{Computer Programming}}, Volume 2: Seminumerical Algorithms},
+  author = {Knuth, Donald Ervin},
+  edition = {3rd},
+  publisher = {{Addison-Wesley}},
+  isbn = {978-0201896848},
+  lccn = {QA76.6 .K64 1997},
+  year = {1997}
+}
+
+@misc{SciMLOrdinaryDiffEqJl,
+  author = {{SciML}},
+  title = {{OrdinaryDiffEq.jl}},
[…]
