### 2023-11-07 #135 @bennn
@lucaferranti it still fails with an empty bib. I think we're blocked here until `editorial-bot` arrives.
