### 2024-02-13 #135 @lucaferranti
Hi @bennn  and @ashton314 , have you noticed this issue? https://github.com/JuliaCon/JuliaConSubmission.jl/issues/34

it seems having underscore in the DOI may cause issues, maybe that's what broke compilation?

(side note, the infra update should be wrapped up within this month, so this should go forward pretty soon in a way or another)
