### 2024-02-15 #135 @bennn
Hi @lucaferranti , no I hadn't seen that, and we do have an underscore (an escaped one but still).

But, we had the same issue with an empty bib of zero references.

I think the intra update is what we need. Very excited for it!
