### 2024-03-24 #135 @lucaferranti
I should maybe also point out that this submission has a bit of a peculiar structure, in the sense that the paper is not a `paper/` folder in a software repository, but its own repository, so `paper.tex` etc. are in the base directory, maybe this breaks the workflow? I still don't understand why it doesn't post a comment about the failure, like in other issues. Looking at the actions: https://github.com/JuliaCon/proceedings-papers/actions it seems for this one it does not even start the generation
