### 2024-03-24 #135 @arfon
@lucaferranti - it looks like the migration of all of the review issues only happened for the ones under active review (i.e., not pre-reviews) so I think we need @xuanxu to re-run his script for the issues like this one as the metadata structure at the top of the issue is currently incorrect.
