### 2024-04-09 #135 @lucaferranti
@ashton314 new infra ready and paper compilation working :tada:

It seems the paper is missing hte references, could you look into it
