### 2024-04-09 #135 @lucaferranti
Hi @oscardssmith 👋 ,

would you be interested in reviewing this submission for the JuliaCon proceedings? You can find reviewer instructions [here](https://juliacon.github.io/proceedings-guide/reviewer/) and you are of course also welcome to ask any questions you may have.
