### 2024-04-10 #135 @JeffreySarnoff
This is nice work, and much appreciated.  With regard to "8-bit formats are on the rise", please consider adding this reference:
[P3109 Interim Report](https://github.com/P3109/Public/blob/main/Shared%20Reports/P3109%20WG%20Interim%20Report.pdf).  That file will continue to be updated periodically, while the link remains the same.  [Here](https://sagroups.ieee.org/p3109wgpublic/) is additional information on IEEE P3109.
