### 2024-04-11 #135 @lucaferranti
@JeffreySarnoff , does this mean you are interested in reviewing this submission?

This issue is at the moment a pre-review, once I find two reviewers, I will start a review issue, where reviewers will have a checklist of things to check and can leave feedback.

Reviewing for JuliaCon proceedings is in general more lightweight (and enjoyable) than normal academic reviews :)
