+++ paper.tex
@@ -106,7 +106,7 @@ from ocean simulations to heat modes~(\cref{s:casestudies}):
 \Fp{} numbers use a finite number of bits to represent a spectrum of points along the real number line~(\cref{f:real-vs-fp}).
 The implementation strategy is essentially that followed in scientific notation.
 A \fp{} number packs a sign bit, an exponent, and a fraction part (also called the ``significand'' or ``mantissa'') into a bitstring.
-Typical strings are 64 or 32 bits long, but 16-bit and 8-bit formats are on the rise~\cite{klowerLowprecisionClimateComputing2021,fp8}.
+Typical strings are 64 or 32 bits long, but 16-bit and 8-bit formats are on the rise~\cite{klowerLowprecisionClimateComputing2021,fp8,p3109}.
 This representation supports very small and very large numbers in a narrow range of bits:
 
 \begin{lstlisting}[language = Julia]
