+++ ref.bib
@@ -367,3 +367,12 @@
   doi          = {10.48550/arXiv.2209.05433},
   year         = {2022}
 }
+
+@misc{p3109,
+  title={{IEEE} Working Group {P3109} Interim Report on 8-bit Binary Floating-point Formats},
+  author={{P3109}},
+  note={\url{https://github.com/P3109/Public/blob/main/Shared%20Reports/P3109%20WG%20Interim%20Report.pdf}, Accessed 2024-04-16},
+  year={2024}
+}
+
+
