### 2024-04-25 #148 @lucaferranti
@jeffreysarnoff, @dpsanders (cc @ashton314) thank you for volunteering as reviewers! :pray: I'll be the editor handling this submission and you can ask me questions any time.

For start, you can generate your checklist by commenting

(do not include other text in the message where you run that command.

**Note**: In this submission, the paper is in its own repository instead of being in the software repository. This is allowed. You can find the software here: https://github.com/utahplt/FloatTracker.jl

As you go through the checklist, you can leave your review comments either as issues in the paper/software repository or directly here. If you have any questions, ping me
