### 2024-04-26 #148 @JeffreySarnoff
Is there an approved way to leave notes.  For example, while there is
no section
titled 'Statement of need'  The Abstract does cover the need.  I do not
know if this suffice; it does seem appropriate to mention.

On Fri, Apr 26, 2024 at 3:09 AM The Open Journals editorial robot <
***@***.***> wrote:

> 👉📄 Download article proof
> <https://raw.githubusercontent.com/JuliaCon/proceedings-papers/jcon.00148/jcon.00148/10.21105.jcon.00148.pdf>
> 📄 View article proof on GitHub
> <https://github.com/JuliaCon/proceedings-papers/blob/jcon.00148/jcon.00148/10.21105.jcon.00148.pdf>
> 📄 👈
>
