### 2024-04-26 #148 @lucaferranti
You can add notes as comment in this thread. When I've reviewed, I generally collect all my notes/feedback in a single comment which I post in the review issue (this one). If you want to recommend changes in the paper / code you can also directly open issues in the repositories. In that case, please do mention this issue also there.

For your question of the statement of need. It is not strictly necassary to have a section with that name, but there should be a clear motivation for the software and the problem it is trying to solve. If you feel this is well described in the abstract/introduction, then you can tick the box
