### 2024-05-08 #148 @JeffreySarnoff
I am quizzical.  The checklist is written for papers that are all about a Julia repository.  FlowFPX is the "hook" yet there is no FlowFPX.jl afaik.  I do see that this is a toolkit, still, having FlowFPX.jl that has the individual tools as [deps] (or subsets) along with a place for the docs to live together would be nice.  Meanwhile -- I don't know what to do with the unchecked boxes .. they seem not applicable or to be considered at a later evolution.  Nonetheless I appreciate the capabilities you have made available.
