### 2024-05-28 #148 @bennn
The best we can do now is a loose coupling, basically with a readme. Putting at all together in a robust-cross-platform Julia package is tough because the different components of FlowFPX use different languages & hardware (C for CSTGs, Nvidia GPUs for GPUFPX).
