### 2024-07-04 #148 @dpsanders
Here are some observations from my review. Apologies for the delay.
I consider that a few of these comments are blocking for the publication of the paper in the JuliaCon proceedings. Others are just general suggestions for improving the package!

In general, this is a nice idea and a very nice package, thanks! :)

- As Jeffrey Sarnoff already mentioned, it’s not too clear what “`FlowFPX`” is, or what a “toolkit” is, and what the relationship of `FloatTracker.jl` is to that.

- I only just understood where the abbreviation “FPX” comes from; this should be made explicit in the paper.

- Consider renaming the package to `TrackedFloats.jl`, since it defines the `TrackedFloat` type. This would be more standard in the Julia ecosystem.

- CSTG has no software license.

- Consider rewriting CSTG in Julia...

- You can use `floatmax(Float64)` instead of `prevfloat(typemax(Float64))` (in the paper).

- Use a parametrized type `TrackedFloat{T <: AbstractFloat}`
  or just `Tracked{T}`.

- Consider allowing to track `Int`s for overflow and underflow.

- Allow to write logs to an in-memory array instead of a file

- Think about thread-safety for the logging

- `—odds::Int64 inject if rand(odds) == 1.` I don’t understand this. How is it used?

- “Since FloatTracker instruments Julia programs, it cannot help directly [with GPUs]."
Conceivably there could be a way to use FloatTracker with Julia’s GPU ecosystem (`CUDA.jl` et al.), but the way that logging works would need to be rethought. This seems like it would be a big project thoug!

- In the paper, refer to package names with their `.jl` termination, e.g. `FloatTracker.jl` and `ShallowWaters.jl`, since this is part of the package name.

- How injection works does not seem well explained.

- “For GPUs it relies on GPU-FPX, through the connection between these tools is loose.”  There doesn’t seem to be any connection at all? :)

- There don’t seem to be any guidelines for contributing to the package.

- After looking at the code it seems like there is an in-memory `LogBuffer`, but I don’t see any docs on how to use it.

- Currently all docs are in the `README`. Consider having actual structured docs.

- There is a lack of consistency in the function names: some start with `ft`, while others do not, such as `config_logger`.
I suggest instead *not* exporting these functions, so they would then be called  as
```
FloatTracker.config(…)
FloatTracker.flush_logs()
```
etc.
This would be more standard for the Julia ecosystem.

- There don’t seem to be API docs.

- To use CSTG it seems like a command-line program needs to be called. This should be callable as a function from the Julia package.

- Ideally the CSTG package itself would be provided as a package via BinaryBuilder (or rewritten in Julia, as suggested above :) )
