### 2024-07-08 #148 @bennn
Thanks @dpsanders ! We'll push an update this month that incorporates review comments.

@lucaferranti would you clarify what you need to see in a revised paper? What are the blockers from your perspective --- we'd like our next revision to be something you and the reviewers can quickly approve.
