### 2024-07-09 #148 @lucaferranti
Thank you @dpsanders and @JeffreySarnoff for the review!

@bennn , the JuliaCon proceedings review is by nature an iterative process, unfortunately I cannot give a strict list of necessary and sufficient conditions to guarantee that next reviewers will approve quickly. If you are unclear, I suggest that you ping the reviewers and comment up-front what changes you are going to do and what changes you are leaving for future work. If you are not planning to not address some changes, it would be good to give a clear motivation for the reasoning behind it. Note, however, that the reviewers may have replies to your rebuttal.

To summarize, if in doubt, discuss with the reviewers first :)
