+++ paper.tex
@@ -230,7 +230,7 @@ FlowFPX can help.
 \section{\FlowFPX{}}
 \label{s:flowfpx}
 
-\FlowFPX{} is a toolkit for tracking down \fp{} exceptions.
+\FlowFPX{} (FPX: Floating Point eXeption) is a toolkit for tracking down \fp{} exceptions.
 The primary tools in this paper are \FT{}, which records lifetimes
 and enables fuzzing, and \CSTG{}s (more precisely, CSTGs), which visualize
 the flow of exceptions.
@@ -851,5 +851,5 @@ and DOE ASCR Award Number DE-SC0022252 and NSF CISE Awards 1956106 and 21241.
 
 % Local Variables:
 % jinx-local-words: "BinFPE CFL CSTG's CSTGs Courant Demmel FPChecker Friedrichs Herbie JuliaHub Klöwer Lewy OrdinaryDiffEq RxInfer ShallowWaters Sherlogs TrackedFloat isnan"
-% citar-bibliography: ("./ref.bib")
+% citar-bibliography: nil
 % End:
