### 2024-07-25 utahplt/juliacon2023-paper#4 @ashton314
Paper edits for Juliacon
Source: https://github.com/JuliaCon/proceedings-review/issues/148

  * [X]  clarify what the FlowFPX toolkit is and its relationship to FloatTracker.jl ("TF.jl" for short)
  * [ ]  rename FloatTracker.jl to TrackedFloats.jl
  * [X]  explicitly say there is no formal link between TF.jl and GPU-FPX
  * [X]  spell out FPX = Floating Point eXception, somewhere
  * [X]  clarify the explanation of how injection works, explain how `odds` is used
  * [X]  fix the `floatmax` issue
  * [X]  acknowledge the threat with thread safety
