### 2024-07-25 993bd73
Fix floatmin/floatmax issue
