### 2024-07-25 bd472be
Typos; attempt to explain injection better; clarify what FlowFPX is
