+++ paper.tex
@@ -80,18 +80,22 @@ related to \NaN{} or \Inf{} exceptions~\cite{github-issues}.
 This paper introduces \FlowFPX{}, a toolkit for debugging
 \fp{} exceptions~(\cref{s:flowfpx})
 that has helped improve a variety of applications,
-from ocean simulations to heat modes~(\cref{s:casestudies}):
-\begin{itemize}
+from ocean simulations to heat modes~(\cref{s:casestudies}).
+\FlowFPX{} consists of three tools that can work together to debug \fp{} computations:
+\begin{enumerate}
   \item
     The centerpiece of \FlowFPX{} is \FT{}, a dynamic analysis tool that
-    selectively monitors for exceptions and fuzzes code for vulnerabilities.
+    selectively monitors for exceptions and fuzzes Julia code for vulnerabilities.
+    \FT{} is a Julia library that operates on Julia code.
   \item
-    To visualize results, \FT{} adapts coalesced stack-trace graphs (CSTGs, or
+    To visualize results, \FlowFPX{} uses coalesced stack-trace graphs (CSTGs, or
     \CSTG{}s)~\cite{humphreySystematicDebuggingMethods2014}
     to summarize the program contexts that handled exceptional values.
+    \CSTG{} is a generic stack-trace graph summarization tool.
   \item
     A companion tool \GPUFPX{}~\cite{llsflg-hpdc-2023} provides fine-grained insights for GPU kernels.
-\end{itemize}
+    \GPUFPX{} is a dynamic library that programmers of GPU kernels can use.
+\end{enumerate}
 
 \noindent%
 The toolkit is online: \url{https://utahplt.github.io/flowfpx}
@@ -108,7 +112,7 @@ The toolkit is online: \url{https://utahplt.github.io/flowfpx}
 
 \Fp{} numbers use a finite number of bits to represent a spectrum of points along the real number line~(\cref{f:real-vs-fp}).
 The implementation strategy is essentially that followed in scientific notation.
-A \fp{} number packs a sign bit, an exponent, and a fraction part (also called the ``significand'' or ``mantissa'') into a bitstring.
+A \fp{} number packs a sign bit, an exponent, and a fraction part (also called the ``significand'' or ``mantissa'') into a bit string.
 Typical strings are 64 or 32 bits long, but 16-bit and 8-bit formats are on the rise~\cite{klowerLowprecisionClimateComputing2021,fp8,p3109}.
 This representation supports very small and very large numbers in a narrow range of bits:
 
@@ -231,18 +235,19 @@ FlowFPX can help.
 \label{s:flowfpx}
 
 \FlowFPX{} (FPX: Floating Point eXeption) is a toolkit for tracking down \fp{} exceptions.
-The primary tools in this paper are \FT{}, which records lifetimes
+The primary tools in the \FlowFPX{} toolkit are \FT{}, which records lifetimes
 and enables fuzzing, and \CSTG{}s (more precisely, CSTGs), which visualize
 the flow of exceptions.
 \GPUFPX{} is a third component of \FlowFPX{} that tracks \fp{} exceptions inside GPU kernels~\cite{llsflg-hpdc-2023}.
+This section covers each tool in detail.
 
 \subsection{\FT{}}
 \label{s:floattracker}
 
-\FT{} tracks exceptional values across their \genpropkill{} lifetime
-and can fuzz code for vulnerabilities by injecting a \Nan{} or \Inf{}
-as the result of an operation.
-\FT{} is implemented in Julia and is available on JuliaHub:
+\FT{} tracks exceptional values in Julia programs across
+their \genpropkill{} lifetime and can fuzz code for vulnerabilities by
[…]
