### 2024-07-25 29d469b
Better description of how injection gets decided
