### 2024-07-26 5254138
Clarify relation between GPU-FPX and FloatTracker (no formal one)
