### 2024-08-01 e7e5b99
jcon2: all pkg names --> +.jl
