+++ paper.tex
@@ -37,10 +37,16 @@
 \newcommand{\code}[1]{\texttt{#1}}
 \newcommand{\FlowFPX}{FlowFPX}
 \newcommand{\GPUFPX}{GPU-FPX}
-\newcommand{\FloatTracker}{FloatTracker}
-\newcommand{\FT}{\FloatTracker}
-\newcommand{\Fp}{Floating-point} % hyphen or no?
-\newcommand{\fp}{floating-point} % hyphen or no?
+\newcommand{\juliapkg}[1]{\texttt{#1.jl}}
+\newcommand{\TrackedFloats}{\juliapkg{TrackedFloats}}
+\newcommand{\TF}{\TrackedFloats}
+\newcommand{\ShallowWaters}{\juliapkg{ShallowWaters}}
+\newcommand{\OrdinaryDiffEq}{\juliapkg{OrdinaryDiffEq}}
+\newcommand{\RxInfer}{\juliapkg{RxInfer}}
+\newcommand{\Sherlogs}{\juliapkg{Sherlogs}}
+\newcommand{\NBodySimulator}{\juliapkg{NBodySimulator}}
+\newcommand{\Fp}{Floating-point}
+\newcommand{\fp}{floating-point}
 \newcommand{\CSTG}{stack graph}
 \newcommand{\CPP}{\code{C++}}
 \newcommand{\Dendro}{\textsc{Dendro}}
@@ -84,9 +90,9 @@ from ocean simulations to heat modes~(\cref{s:casestudies}).
 \FlowFPX{} consists of three tools that can work together to debug \fp{} computations:
 \begin{enumerate}
   \item
-    The centerpiece of \FlowFPX{} is \FT{}, a dynamic analysis tool that
+    The centerpiece of \FlowFPX{} is \TF{}, a dynamic analysis tool that
     selectively monitors for exceptions and fuzzes Julia code for vulnerabilities.
-    \FT{} is a Julia library that operates on Julia code.
+    \TF{} is a Julia library that operates on Julia code.
   \item
     To visualize results, \FlowFPX{} uses coalesced stack trace graphs (CSTGs, or
     \CSTG{}s)~\cite{humphreySystematicDebuggingMethods2014}
@@ -235,18 +241,18 @@ FlowFPX can help.
 \label{s:flowfpx}
 
 \FlowFPX{} (FPX: Floating Point eXeption) is a toolkit for tracking down \fp{} exceptions.
-The primary tools in the \FlowFPX{} toolkit are \FT{}, which records lifetimes
+The primary tools in the \FlowFPX{} toolkit are \TF{}, which records lifetimes
 and enables fuzzing, and \CSTG{}s (more precisely, CSTGs), which visualize
 the flow of exceptions.
 \GPUFPX{} is a third component of \FlowFPX{} that tracks \fp{} exceptions inside GPU kernels~\cite{llsflg-hpdc-2023}.
 This section covers each tool in detail.
 
-\subsection{\FT{}}
+\subsection{\TF{}}
 \label{s:floattracker}
 
-\FT{} tracks exceptional values in Julia programs across
+\TF{} tracks exceptional values in Julia programs across
 their \genpropkill{} lifetime and can fuzz code for vulnerabilities by
-injecting a \Nan{} or \Inf{} as the result of an operation. \FT{} is
+injecting a \Nan{} or \Inf{} as the result of an operation. \TF{} is
 implemented in Julia as a library and is available on JuliaHub:
 
 \noindent\begin{center}\noindent\!\!\url{https://juliahub.com/ui/Packages/FloatTracker/dBXig}\!\!
@@ -266,7 +272,7 @@ implemented in Julia as a library and is available on JuliaHub:
   \label{fig:lifetime-class}
 \end{figure}
 
[…]
