### 2024-08-01 5dd2918
jcon2: move future-work threads issue to the future work section
