+++ paper.tex
@@ -75,7 +75,7 @@ unreliable~(\cref{s:background}).
 Exceptional values, such as Not a Number (\Nan{}) and infinity (\Inf{}),
 can and often do arise thanks to culprits such as roundoff error,
 catastrophic cancellation, singular matrices, and vanishing
-derivatives~\cite{sdjmrstp-pc-2022,ddghlllprr-correctness-2022,gllprt-correctness-2021,fpchecker-reports,llg-soap-2022,bllmg-xloop-2022}.
+derivatives~\cite{bllmg-xloop-2022,ddghlllprr-correctness-2022,fpchecker-reports,gllprt-correctness-2021,llg-soap-2022,sdjmrstp-pc-2022}.
 Developers are responsible for guarding against exceptions, but this task
 is difficult because many operations can generate and propagate exceptions.
 Worst of all, operations such as less-than (\code{<}) can kill an exceptional value,
@@ -91,17 +91,16 @@ from ocean simulations to heat modes~(\cref{s:casestudies}).
 \FlowFPX{} consists of three tools that can work together to debug \fp{} computations:
 \begin{enumerate}
   \item
-    The centerpiece of \FlowFPX{} is \TF{}, a dynamic analysis tool that
-    selectively monitors for exceptions and fuzzes Julia code for vulnerabilities.
-    \TF{} is a Julia library that operates on Julia code.
+    The centerpiece of \FlowFPX{} is \TF{}, a dynamic analysis tool written in Julia
+    for Julia code that monitors exceptions and fuzzes for vulnerabilities.
   \item
     To visualize results, \FlowFPX{} uses coalesced stack trace graphs (CSTGs, or
-    \CSTG{}s)~\cite{humphreySystematicDebuggingMethods2014}
-    to summarize the program contexts that handled exceptional values.
-    \CSTG{} is a generic stack trace graph summarization tool.
+    \CSTG{}s)~\cite{humphreySystematicDebuggingMethods2014}.
+    Our CSTG library is written in \CPP{} and accepts text-format stack traces as input.
   \item
-    A companion tool \GPUFPX{}~\cite{llsflg-hpdc-2023} provides fine-grained insights for GPU kernels.
-    \GPUFPX{} is a dynamic library that programmers of GPU kernels can use.
+    For programs that offload work to CUDA GPUs,
+    the binary instrumentation tool \GPUFPX{}~\cite{llsflg-hpdc-2023}
+    adds logging to kernels.
 \end{enumerate}
 
 \noindent%
@@ -132,8 +131,8 @@ julia> floatmin(Float64)
 
 The flip side is that most real numbers fall into the gaps between \fp{} numbers and must be rounded, which introduces error
 and can lead to surprising results.
-For example, adding the tiny Planck constant to the large Avogadro number results
-in Avogadro's number after rounding:
+For example, adding the tiny Planck constant to the large Avogadro number results in
+the Avogadro number after rounding:
 
 \begin{lstlisting}[language = Julia]
 julia> planck = 6.62607015e-34
@@ -175,11 +174,11 @@ NaN
 IEEE~754 defers the question of how to handle such exceptions to application code.
 It is up to developers to watch for \Nan{}, \Inf{}, and subnormal numbers (underflow)
 and implement an appropriate repair.
-This is no easy task, however, because of the approximations inherent to
+This is a difficult task because of the approximations inherent to
 floating point.
 Even an apparently-safe division could result in a \Nan{} if its
 denominator gets truncated to zero.
-Some \Nan{}s might be spurious, others might be fatal, but in any event
+Some \Nan{}s might be spurious while others might be fatal, but in any event
 anticipating the various exceptions is a burden.
 All too often, exceptional values go unhandled and flow through the code.
 
@@ -241,7 +240,7 @@ FlowFPX can help.
[…]
