### 2024-08-02 c12796f
jcon2: number → constant
