+++ paper.tex
@@ -131,8 +131,8 @@ julia> floatmin(Float64)
 
 The flip side is that most real numbers fall into the gaps between \fp{} numbers and must be rounded, which introduces error
 and can lead to surprising results.
-For example, adding the tiny Planck constant to the large Avogadro number results in
-the Avogadro number after rounding:
+For example, adding the tiny Planck constant to the large Avogadro constant results in
+the Avogadro constant after rounding:
 
 \begin{lstlisting}[language = Julia]
 julia> planck = 6.62607015e-34
