### 2024-08-02 #148 @ashton314
Status update: lots of concerns addressed; currently waiting for registering the package rename to TrackedFloats on JuliaHub: https://github.com/JuliaRegistries/General/pull/112323
