### 2024-08-08 #148 @bennn
@JeffreySarnoff @dpsanders  the latest PDF has fixes for everything except the juliahub package URL in section 3.
