### 2024-08-08 5594dd4
jcon2: update JuliaHub URL
