+++ paper.tex
@@ -253,9 +253,9 @@ This section covers each tool in detail.
 \TF{} tracks exceptional values in Julia programs across
 their \genpropkill{} lifetime and can fuzz code for vulnerabilities by
 injecting a \Nan{} or \Inf{} as the result of an operation. \TF{} is
-implemented in Julia as a library and is available on JuliaHub (TODO update link):
+implemented in Julia as a library and is available on JuliaHub:
 
-\noindent\begin{center}\noindent\!\!\url{https://juliahub.com/ui/Packages/FloatTracker/dBXig}\!\!
+\noindent\begin{center}\noindent\!\!\url{https://juliahub.com/ui/Packages/TrackedFloats}\!\!
 \end{center}
 
 \subsubsection{Tracking Exceptional Values}
