### 2024-08-08 #148 @ashton314
How extensive must the contribution guidelines be? I've added a section to our README advising bug reporters to report issues on GitHub and to open PRs there if they have code they'd like to contribute.
