### 2024-08-08 #148 @JeffreySarnoff
While I don't know the answer to your question, referring to [this guidance from SciML](https://github.com/SciML/ColPrac) or copying what may be most salient for you is one way to go.
