### 2024-08-08 #148 @JeffreySarnoff
by the way, the message "⚠️ NOTICE ⚠️ we are in the process of renaming FloatTracker → TrackedFloats to bring this inline with Julia package naming conventions. Please be patient." at the top of the README for TrackedFloats.jl is not useful any longer (imo).
