### 2024-08-28 #148 @ashton314
@lucaferranti I'm gearing up for a talk at ECOOP in September. I hope to make a little progress before then, but it might not be finished until after the conference.
