### 2024-08-30 #148 @ashton314
Timeline update: today I've got a little time and I'll try making sure everything exported gets a `tf_` and the beginning.

Could I get some suggestions on what constitutes "adding structure to the documentation"? Moreover—this isn't a big package; I don't think the README is that long and it serves the need for documentation adequately. Is that really going to be a blocker?
