### 2024-08-30 #148 @ashton314
So, concretely: I have a little more work to do today on the software; depending on how we feel about the state of documentation, it might take me until late September when I have some time after ECOOP.
