### 2024-08-31 #148 @lucaferranti
@ashton314 the timeline sounds good, thanks for the updates.

For your question, I would understand it as setting up Documenter.jl and have an API documentation by collecting exported functions docstrings as well as a couple of tutorials, but do feel free to ask @dpsanders directly to clarify what he meant if it's not clear to you
