### 2024-10-15 #148 @bennn
@lucaferranti @dpsanders @JeffreySarnoff  we're done on our end. Waiting now on your reviews.

All the TODOs [above](https://github.com/JuliaCon/proceedings-review/issues/148#issuecomment-2222008483) are either complete or have links to the [Tracked Floats issue tracker](https://github.com/utahplt/TrackedFloats.jl/issues).
