### 2024-10-15 #148 @JeffreySarnoff
I may have missed this in the paper -- in Figure 6 there are numbered arrows, it is unclear what those numbers convey.
