### 2024-10-16 #148 @ashton314
> I may have missed this in the paper -- in Figure 6 there are numbered arrows, it is unclear what those numbers convey.

This is explained in §3.2, paragraph 2.
