### 2024-12-02 #148 @lucaferranti
I'll try to reach david via slack
