### 2024-12-05 #148 @JeffreySarnoff
In the event that David is unavailable, you should proceed.
