### 2024-12-06 #148 @dpsanders
Hi, sorry for the delay.

A few comments:
- The link at https://utahplt.github.io/flowfpx/ to the Julia package has the old name.
- This is probably more a problem with the style file, but copying and pasting code from the PDF to Julia does not work properly.
- I would link to the package page on GitHub rather than JuliaHub from the paper.
- The case studies are very cool!
- The paper does not have a section titled "Statement of need"; however, I have checked the box since the need is clear from the paper.

Overall, this is very nice work, thanks!
I recommend to proceed with publication.
