### 2024-12-10 #148 @bennn
Thanks David!

We fixed the name at https://utahplt.github.io/flowfpx/ . We didn't change the JuliaCon style file but agree it's a bummer we can't copy/paste. We also didn't change the link in the paper because we feel that JuliaHub is the authoritative home for the code. GitHub is incidental ... if we decided to move off it for Microsoft/OpenAI/whatever concerns people should still be able to find TrackedFloats on JuliaHub.
