### 2024-12-10 #148 @lucaferranti
@bennn to finalize the acceptance I need:

- the latest version of the software associated with the paper
- a zenodo DOI, since this submission is effectively regarding the software, would be better to have the software repo archived in zenodo and use the corresponding DOI
