### 2024-12-10 #148 @lucaferranti
@bennn to finalize the acceptance, please see the "additional author tasks after review is complete" in the message above :point_up:

Also, there seems to be an invalid DOI, could you double check that?

```
❌ INVALID DOIs

- 10.1007/978-3-031-08751-6\_9 URL is INVALID
```

(for the missing doi, that seems to point to a reivew of KNUTH book and not the book itself so a false positive)
