### 2024-12-11 #148 @ashton314
DOI for Zenodo archive: https://doi.org/10.5281/zenodo.14396141
Software release version: 2.1.0 [release](https://github.com/utahplt/TrackedFloats.jl/releases/tag/2.1.0)
