### 2024-12-11 #148 @ashton314
@lucaferranti I've made a Zenodo archive and a release for TrackedFloats

The DOI that is "missing" doesn't get recognized because I have to escape an underscore in the `.bib` file. Normally you shouldn't have to do this, but something's broken.
