### 2024-12-12 #148 @lucaferranti
ok there is a 2.1.0 tag, but that does not corresponds to the software version it seems?
