### 2024-12-12 #148 @ashton314
@lucaferranti I fixed the version in Project.toml; I had to bump the tag to 2.1.1. [Release here](https://github.com/utahplt/TrackedFloats.jl/releases/tag/2.1.1)
