### 2024-12-12 #148 @lucaferranti
Great! Thank you for fixing it. Did you register the version 2.1.0 ? `add TrackedFloats` installs version 2.0.0
