### 2024-12-21 #148 @ashton314
Is there a way to ask JuliaHub to pull the most recent version from the UI? Where are the docs on this? I can only find the docs for submitting a *new* package.
