### 2024-12-21 #148 @lucaferranti
> Is there a way to ask JuliaHub to pull the most recent version from the UI? Where are the docs on this? I can only find the docs for submitting a new package.

 I use [JuliaRegistrator](https://github.com/JuliaRegistries/Registrator.jl) via github to manage package versions, to make a new release, I just comment `@JuliaRegistrator register` on a commit and that releases the new version (creates PR to general registry) and makes the tag. I've actually never used the juliahub web interface, so unfortunately I don't know the answer to that.

Maybe ask on the julia slack? `helpdesk` or `pkg-registration` are probably good channels for that
