### 2025-01-07 #148 @lucaferranti
Hi @ashton314 :wave: ,

any luck with getting the new version released? this is the very last step before getting this finally published :)
