### 2025-01-07 #148 @ashton314
I think it merged! https://github.com/JuliaRegistries/General/pull/122091

I don't see the update on JuliaHub just yet… I expect it takes a minute.
