### 2025-01-07 #148 @ashton314
@lucaferranti @bennn The new version is up on JuliaHub! :shipit:
