### 2025-01-08 #148 @lucaferranti
@ashton314 we should be ready to go. Please have a final look at the paper pdf and if/when everything looks good, ping me and I'll complete the acceptance.
